[CAES-rebuild-store] Splitted logic from node to list, item, childItem
