+++ common/constants.js
@@ -57,6 +57,11 @@ export const LIST_MODES = {
 export const ITEM_CREDENTIALS_TYPE = 'credentials';
 export const ITEM_CREDIT_CARD_TYPE = 'card';
 export const ITEM_DOCUMENT_TYPE = 'document';
+export const ITEM_TYPES_ARRAY = [
+  ITEM_CREDENTIALS_TYPE,
+  ITEM_CREDIT_CARD_TYPE,
+  ITEM_DOCUMENT_TYPE,
+];
 export const ITEM_TYPES = {
   ITEM_CREDENTIALS_TYPE,
   ITEM_CREDIT_CARD_TYPE,
