+++ common/normalizers/schemas/member.js
@@ -1,0 +1,5 @@
+import { schema } from 'normalizr';
+
+const memberSchema = new schema.Entity('byId');
+
+export default memberSchema;
