+++ common/reducers/index.js
@@ -1,3 +1,6 @@
-export { default as nodeReducer } from './node';
 export { default as memberReducer } from './member';
 export { default as userReducer } from './user';
+export { default as workflowReducer } from './workflow';
+export { default as listReducer } from './list';
+export { default as itemReducer } from './item';
+export { default as childItemReducer } from './childItem';
