+++ containers/Bootstrap/index.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 import { setMasterPassword } from 'common/actions/user';
-import { resetStore } from 'common/actions/node';
+import { resetStore } from 'common/actions/workflow';
 import { masterPasswordSelector } from 'common/selectors/user';
 import Bootstrap from './Bootstrap';
