+++ containers/Import/index.js
@@ -1,1 +1,27 @@
-export { default as Import } from './Import';
+import { connect } from 'react-redux';
+import { createItemsBatchRequest } from 'common/actions/item';
+import { fetchNodesRequest } from 'common/actions/workflow';
+import { fetchKeyPairRequest, fetchUserSelfRequest } from 'common/actions/user';
+import { createStructuredSelector } from 'reselect';
+import { selectableListsWithoutChildrenSelector } from 'common/selectors/list';
+import { keyPairSelector } from 'common/selectors/user';
+import { isLoadingSelector } from 'common/selectors/workflow';
+import Import from './Import';
+
+const mapStateToProps = createStructuredSelector({
+  lists: selectableListsWithoutChildrenSelector,
+  keyPair: keyPairSelector,
+  isLoading: isLoadingSelector,
+});
+
+const mapDispatchToProps = {
+  fetchNodesRequest,
+  fetchKeyPairRequest,
+  fetchUserSelfRequest,
+  createItemsBatchRequest,
+};
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(Import);
