+++ containers/index.js
@@ -3,7 +3,7 @@ export { default as DashboardContainer } from './Dashboard';
 export { default as ManageListContainer } from './ManageList';
 export { Sharing } from './Sharing';
 export { Invite } from './Invite';
-export { Import } from './Import';
+export { default as ImportContainer } from './Import';
 export { SignInContainer } from './SignIn';
 export { SignUpContainer } from './SignUp';
 export { ResetPasswordContainer } from './ResetPassword';
